+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 
+## [1.1.0] - 2025-12-05
+
+- Make url input prefix configurable
+- Make border around field group configurable
+- Make embed preview configurable
+
 ## [1.0.0] - 2025-12-02
 
 - Initial release
 
+[1.1.0]: https://github.com/daun/statamic-embed-fieldtype/releases/tag/1.1.0
 [1.0.0]: https://github.com/daun/statamic-embed-fieldtype/releases/tag/1.0.0